Deduplicate LinuxUserDefinedFileAttributeView + BsdUserDefinedFileAttributeView
