+++ src/java.base/macosx/classes/sun/nio/fs/BsdFileSystem.java
@@ -69,6 +69,7 @@ class BsdFileSystem extends UnixFileSystem {
 
     @Override
     void copyNonPosixAttributes(int ofd, int nfd) {
+        UnixUserDefinedFileAttributeView.copyExtendedAttributes(ofd, nfd);
     }
 
     /**
