+++ src/java.base/share/classes/sun/nio/fs/NativeBuffer.java
@@ -33,7 +33,7 @@ import jdk.internal.ref.CleanerFactory;
  * A light-weight buffer in native memory.
  */
 
-class NativeBuffer {
+class NativeBuffer implements AutoCloseable {
     private static final Unsafe unsafe = Unsafe.getUnsafe();
 
     private final long address;
@@ -59,6 +59,11 @@ class NativeBuffer {
         this.size = size;
         this.cleanable = CleanerFactory.cleaner()
                                        .register(this, new Deallocator(address));
+    }
+
+    @Override
+    public void close() {
+        release();
     }
 
     void release() {
