+++ src/java.base/unix/classes/sun/nio/fs/UnixNativeDispatcher.java
@@ -564,6 +564,52 @@ class UnixNativeDispatcher {
     static native byte[] strerror(int errnum);
 
     /**
+     * ssize_t fgetxattr(int filedes, const char *name, void *value, size_t size);
+     */
+    static int fgetxattr(int filedes, byte[] name, long valueAddress,
+                         int valueLen) throws UnixException
+    {
+        try (NativeBuffer buffer = NativeBuffers.asNativeBuffer(name)) {
+            return fgetxattr0(filedes, buffer.address(), valueAddress, valueLen);
+        }
+    }
+
+    private static native int fgetxattr0(int filedes, long nameAddress,
+                                         long valueAddress, int valueLen) throws UnixException;
+
+    /**
+     *  fsetxattr(int filedes, const char *name, const void *value, size_t size, int flags);
+     */
+    static void fsetxattr(int filedes, byte[] name, long valueAddress,
+                          int valueLen) throws UnixException
+    {
+        try (NativeBuffer buffer = NativeBuffers.asNativeBuffer(name)) {
+            fsetxattr0(filedes, buffer.address(), valueAddress, valueLen);
+        }
+    }
+
+    private static native void fsetxattr0(int filedes, long nameAddress,
+                                          long valueAddress, int valueLen) throws UnixException;
+
+    /**
+     * fremovexattr(int filedes, const char *name);
+     */
+    static void fremovexattr(int filedes, byte[] name) throws UnixException {
+        try (NativeBuffer buffer = NativeBuffers.asNativeBuffer(name)) {
+            fremovexattr0(filedes, buffer.address());
+        }
+    }
+
+    private static native void fremovexattr0(int filedes, long nameAddress)
+            throws UnixException;
+
+    /**
+     * size_t flistxattr(int filedes, const char *list, size_t size)
+     */
+    static native int flistxattr(int filedes, long listAddress, int size)
+            throws UnixException;
+
+    /**
      * Capabilities
      */
     private static final int SUPPORTS_OPENAT        = 1 << 1;  // syscalls
